login/logout/register button fixes
